药品增删改的接口实现

+++ backend/src/main/java/com/example/petbackend/service/medicine/MedicineService.java
@@ -5,8 +5,8 @@ import java.util.Map;
 
 public interface MedicineService {
 
-    Map<String, String> addMedicine(String medicine_name, Integer medicine_cost);
-    Map<String, String> updateMedicine(Integer medicine_id, Integer medicine_cost);
+    Map<String, String> addMedicine(String medicine_name, Double medicine_cost);
+    Map<String, String> updateMedicine(Integer medicine_id, Double medicine_cost);
     Map<String, String> deleteMedicine(Integer medicine_id);
 
 }
